==> Step1_DownloadStats.ipynb <==
import requests
import json
import time,datetime

sundays=[]
mondays=[]
thursdays=[]
saturdays=[]
for year in [2010]:
    for month in (9,10,11,12,1,2):
        for date in range(1,32):
            try:
                if datetime.date(year,month,date).weekday()==6:
                    sundays.append(str(year)+(str(month)).zfill(2)+(str(date)).zfill(2))
                elif datetime.date(year,month,date).weekday()==0:
                    mondays.append(str(year)+(str(month)).zfill(2)+(str(date)).zfill(2))
                elif datetime.date(year,month,date).weekday()==3:
                    thursdays.append(str(year)+(str(month)).zfill(2)+(str(date)).zfill(2))
                elif datetime.date(year,month,date).weekday()==5:
                    saturdays.append(str(year)+(str(month)).zfill(2)+(str(date)).zfill(2))
                    
            except:
                continue

gameurls={}
for gamedate in sundays+mondays+thursdays+saturdays:
    gameurls[gamedate]=[]
    for _x in range(16):
        if _x<10:
            x='0'+str(_x)
        else:
            x=str(_x)

        url='http://www.nfl.com/liveupdate/game-center/' + gamedate + x + '/' + gamedate + x + '_gtd.json'
        gameurls[gamedate].append(url)

listofgameurls=[eachgameurl for gamedayurls in gameurls.values() for eachgameurl in gamedayurls]
for gameurl in listofgameurls:
    response=requests.get(gameurl)
    if response.status_code==200:
        gamedata=response.json()
        storedata=gamedata[list(gamedata.keys())[0]]
        filenametouse=list(gamedata.keys())[0] + '.txt'
        with open(filenametouse,'w') as filename:
            json.dump(storedata,filename)
        print('File Saved - ',filenametouse)
        time.sleep(30)

gamedata
