==> Step3_CreateTrainingSet.ipynb <==
import pandas as pd
import numpy as np

class Game:
    
    def __init__(self,myteam,oppteam,gameid,bigdf):
        self.myteam=myteam
        self.oppteam=oppteam
        self.gameid=gameid
        self.myteam_allgamesdf=self.getmyteamAllGamesDF(bigdf)
        self.oppteam_allgamesdf=self.getoppteamAllGamesDF(bigdf)
        
    @staticmethod
    def gettime(text):
        return int(text.split(':')[0])+int(text.split(':')[1])/60
        
    def getmyteamAllGamesDF(self,allgamesdf):
        
        df=allgamesdf[(allgamesdf.hometeam==self.myteam) | (allgamesdf.awayteam==self.myteam)]
        return df.sort_values(by='gamedate',ascending=False).reset_index(drop=True)
    
    def getoppteamAllGamesDF(self,allgamesdf):
        
        df=allgamesdf[(allgamesdf.hometeam==self.oppteam) | (allgamesdf.awayteam==self.oppteam)]
        return df.sort_values(by='gamedate',ascending=False).reset_index(drop=True)
    
    def getmyteamPreviousWins(self,lookback_noofgames):

        df=self.myteam_allgamesdf
        df=df.loc[self.gameid+1:self.gameid+lookback_noofgames,:]
        wins=[]
        for index in df.index:
            if df.loc[index,'hometeam']==self.myteam:
                if df.loc[index,'hometeam_score']>df.loc[index,'awayteam_score']:
                    wins.append(1)
                else:
                    wins.append(0)
            elif df.loc[index,'awayteam']==self.myteam:
                if df.loc[index,'awayteam_score']>df.loc[index,'hometeam_score']:
                    wins.append(1)
                else:
                    wins.append(0)
        return wins
    
    def getoppteamPreviousWins(self,lookback_noofgames):
        
        df=self.oppteam_allgamesdf
        df=df.loc[self.gameid+1:self.gameid+lookback_noofgames,:]
        wins=[]
        for index in df.index:
            if df.loc[index,'hometeam']==self.oppteam:
                if df.loc[index,'hometeam_score']>df.loc[index,'awayteam_score']:
                    wins.append(1)
                else:
                    wins.append(0)
            elif df.loc[index,'awayteam']==self.oppteam:
                if df.loc[index,'awayteam_score']>df.loc[index,'hometeam_score']:
                    wins.append(1)
                else:
                    wins.append(0)
        return wins
    
    def getGameDate(self):
        
        return self.myteam_allgamesdf.loc[self.gameid,'gamedate']
    
    
    def getHomeOrAway(self):
        
        return self.myteam_allgamesdf.loc[self.gameid,'hometeam']==self.myteam
    
    def getGameResult(self):
        
        result=0
        df=self.myteam_allgamesdf.loc[self.gameid,:]
        if df['hometeam']==self.myteam:
            if df['hometeam_score']>df['awayteam_score']:
                result=1
            else:
                result=0
        elif df['awayteam']==self.myteam:
            if df['awayteam_score']>df['hometeam_score']:
                result=1
            else:
                result=0
        return result
    
    def getmyteam_CurScore(self):
        
        df=self.myteam_allgamesdf.loc[self.gameid,:]
        if df['hometeam']==self.myteam:
            return df['hometeam_score']
        else:
            return df['awayteam_score']
        
    def getoppteam_CurScore(self):
        
        df=self.myteam_allgamesdf.loc[self.gameid,:]
        if df['hometeam']==self.myteam:
            return df['awayteam_score']
        else:
            return df['hometeam_score']
        
    def getmyteamScores(self,lookback_noofgames):
        
        df=self.myteam_allgamesdf
        df=df.loc[self.gameid+1:self.gameid+lookback_noofgames,:]
        scores=[]
        for index in df.index:
            if df.loc[index,'hometeam']==self.myteam:
                scores.append(df.loc[index,'hometeam_score'])
            elif df.loc[index,'awayteam']==self.myteam:
                scores.append(df.loc[index,'awayteam_score'])
        return scores
    
    
    def getoppteamScores(self,lookback_noofgames):
        
        df=self.oppteam_allgamesdf
        df=df.loc[self.gameid+1:self.gameid+lookback_noofgames,:]
        scores=[]
        for index in df.index:
            if df.loc[index,'hometeam']==self.oppteam:
                scores.append(df.loc[index,'hometeam_score'])
            elif df.loc[index,'awayteam']==self.oppteam:
                scores.append(df.loc[index,'awayteam_score'])
        return scores
    
    def getmyteamTrnOvr(self,lookback_noofgames):
        
        df=self.myteam_allgamesdf
        df=df.loc[self.gameid+1:self.gameid+lookback_noofgames,:]
        trnovrs=[]
        for index in df.index:
            if df.loc[index,'hometeam']==self.myteam:
                trnovrs.append(df.loc[index,'homestats_trnovr'])
            elif df.loc[index,'awayteam']==self.myteam:
                trnovrs.append(df.loc[index,'awaystats_trnovr'])
        return trnovrs
    
    def getoppteamTrnOvr(self,lookback_noofgames):
        
        df=self.oppteam_allgamesdf
        df=df.loc[self.gameid+1:self.gameid+lookback_noofgames,:]
        trnovrs=[]
        for index in df.index:
            if df.loc[index,'hometeam']==self.oppteam:
                trnovrs.append(df.loc[index,'homestats_trnovr'])
            elif df.loc[index,'awayteam']==self.oppteam:
                trnovrs.append(df.loc[index,'awaystats_trnovr'])
        return trnovrs
    
    def getmyteamTop(self,lookback_noofgames):
    
        df=self.myteam_allgamesdf
        df=df.loc[self.gameid+1:self.gameid+lookback_noofgames,:]
        timeofpos=[]
        for index in df.index:
            if df.loc[index,'hometeam']==self.myteam:
                timeofpos.append(Game.gettime(df.loc[index,'homestats_top']))
            elif df.loc[index,'awayteam']==self.myteam:
                timeofpos.append(Game.gettime(df.loc[index,'awaystats_top']))
        return timeofpos
    
    def getoppteamTop(self,lookback_noofgames):
        
        df=self.oppteam_allgamesdf
        df=df.loc[self.gameid+1:self.gameid+lookback_noofgames,:]
        timeofpos=[]
        for index in df.index:
            if df.loc[index,'hometeam']==self.oppteam:
                timeofpos.append(Game.gettime(df.loc[index,'homestats_top']))
            elif df.loc[index,'awayteam']==self.oppteam:
                timeofpos.append(Game.gettime(df.loc[index,'awaystats_top']))
        return timeofpos
    
    def getmyteamYds(self,ydstype,lookback_noofgames):
    
        df=self.myteam_allgamesdf
        df=df.loc[self.gameid+1:self.gameid+lookback_noofgames,:]
        yds=[]
        for index in df.index:
            if ydstype=='rush':
                if df.loc[index,'hometeam']==self.myteam:
                    yds.append(df.loc[index,'homestats_ryds'])
                elif df.loc[index,'awayteam']==self.myteam:
                    yds.append(df.loc[index,'awaystats_ryds'])
            
            elif ydstype=='pass':
                if df.loc[index,'hometeam']==self.myteam:
                    yds.append(df.loc[index,'homestats_pyds'])
                elif df.loc[index,'awayteam']==self.myteam:
                    yds.append(df.loc[index,'awaystats_pyds'])
            
            elif ydstype=='punt':
                if df.loc[index,'hometeam']==self.myteam:
                    yds.append(df.loc[index,'homestats_ptyds'])
                elif df.loc[index,'awayteam']==self.myteam:
                    yds.append(df.loc[index,'awaystats_ptyds'])
                    
            elif ydstype=='pen':
                if df.loc[index,'hometeam']==self.myteam:
                    yds.append(df.loc[index,'homestats_penyds'])
                elif df.loc[index,'awayteam']==self.myteam:
                    yds.append(df.loc[index,'awaystats_penyds'])
        
        return yds
    
    
    def getoppteamYds(self,ydstype,lookback_noofgames):
        
        df=self.oppteam_allgamesdf
        df=df.loc[self.gameid+1:self.gameid+lookback_noofgames,:]
        yds=[]
        for index in df.index:
            if ydstype=='rush':
                if df.loc[index,'hometeam']==self.oppteam:
                    yds.append(df.loc[index,'homestats_ryds'])
                elif df.loc[index,'awayteam']==self.oppteam:
                    yds.append(df.loc[index,'awaystats_ryds'])
            
            elif ydstype=='pass':
                if df.loc[index,'hometeam']==self.oppteam:
                    yds.append(df.loc[index,'homestats_pyds'])
                elif df.loc[index,'awayteam']==self.oppteam:
                    yds.append(df.loc[index,'awaystats_pyds'])
            
            elif ydstype=='punt':
                if df.loc[index,'hometeam']==self.oppteam:
                    yds.append(df.loc[index,'homestats_ptyds'])
                elif df.loc[index,'awayteam']==self.oppteam:
                    yds.append(df.loc[index,'awaystats_ptyds'])
                    
            elif ydstype=='pen':
                if df.loc[index,'hometeam']==self.oppteam:
                    yds.append(df.loc[index,'homestats_penyds'])
                elif df.loc[index,'awayteam']==self.oppteam:
                    yds.append(df.loc[index,'awaystats_penyds'])
        
        return yds
    
    
    
# game=Game('PHI','DAL',3)
# game.getmyteamYds('rush',3),game.getoppteamYds('pass',3)

def GetTrainingDF(lookbacknoofgames,bigdf):
    
    teamlist=list(set(bigdf.hometeam.values))
    
    mytrainingdf={}
    mytrainingdf['myteam']=[]
    mytrainingdf['oppteam']=[]
    mytrainingdf['myteam_prvwins']=[]
    mytrainingdf['oppteam_prvwins']=[]
    mytrainingdf['homegame']=[]
    mytrainingdf['result']=[]
    mytrainingdf['myteam_scores']=[]
    mytrainingdf['oppteam_scores']=[]
    mytrainingdf['myteam_trnovr']=[]
    mytrainingdf['oppteam_trnovr']=[]
    mytrainingdf['myteam_top']=[]
    mytrainingdf['oppteam_top']=[]
    mytrainingdf['myteam_pyds']=[]
    mytrainingdf['oppteam_pyds']=[]
    mytrainingdf['myteam_ryds']=[]
    mytrainingdf['oppteam_ryds']=[]
    mytrainingdf['myteam_penyds']=[]
    mytrainingdf['oppteam_penyds']=[]
    mytrainingdf['myteam_ptyds']=[]
    mytrainingdf['oppteam_ptyds']=[]
    mytrainingdf['gamedate']=[]
    mytrainingdf['myteam_curscore']=[]
    mytrainingdf['oppteam_curscore']=[]


    for team in teamlist:
        df=bigdf[(bigdf.hometeam==team) | (bigdf.awayteam==team)].sort_values(by='gamedate',ascending=False).reset_index(drop=True)
        oppteams=[df.loc[index,'hometeam'] if df.loc[index,'hometeam']!=team else df.loc[index,'awayteam'] for index in df.index ]
        gameids=df.index

        for gameid,myteam,oppteam in list(zip(gameids,[team]*len(oppteams),oppteams))[:-lookbacknoofgames]:
            game=Game(myteam,oppteam,gameid,bigdf)
            mytrainingdf['myteam'].append(game.myteam)
            mytrainingdf['oppteam'].append(game.oppteam)
            mytrainingdf['myteam_prvwins'].append(game.getmyteamPreviousWins(lookbacknoofgames))
            mytrainingdf['oppteam_prvwins'].append(game.getoppteamPreviousWins(lookbacknoofgames))
            mytrainingdf['gamedate'].append(game.getGameDate())
            mytrainingdf['homegame'].append(game.getHomeOrAway())
            mytrainingdf['result'].append(game.getGameResult())
            mytrainingdf['myteam_scores'].append(game.getmyteamScores(lookbacknoofgames))
            mytrainingdf['oppteam_scores'].append(game.getoppteamScores(lookbacknoofgames))
            mytrainingdf['myteam_trnovr'].append(game.getmyteamTrnOvr(lookbacknoofgames))
            mytrainingdf['oppteam_trnovr'].append(game.getoppteamTrnOvr(lookbacknoofgames))
            mytrainingdf['myteam_top'].append(game.getmyteamTop(lookbacknoofgames))
            mytrainingdf['oppteam_top'].append(game.getoppteamTop(lookbacknoofgames))
            mytrainingdf['myteam_pyds'].append(game.getmyteamYds('pass',lookbacknoofgames))
            mytrainingdf['oppteam_pyds'].append(game.getoppteamYds('pass',lookbacknoofgames))
            mytrainingdf['myteam_ryds'].append(game.getmyteamYds('rush',lookbacknoofgames))
            mytrainingdf['oppteam_ryds'].append(game.getoppteamYds('rush',lookbacknoofgames))
            mytrainingdf['myteam_penyds'].append(game.getmyteamYds('pen',lookbacknoofgames))
            mytrainingdf['oppteam_penyds'].append(game.getoppteamYds('pen',lookbacknoofgames))
            mytrainingdf['myteam_ptyds'].append(game.getmyteamYds('punt',lookbacknoofgames))
            mytrainingdf['oppteam_ptyds'].append(game.getoppteamYds('punt',lookbacknoofgames))
            mytrainingdf['myteam_curscore'].append(game.getmyteam_CurScore())
            mytrainingdf['oppteam_curscore'].append(game.getoppteam_CurScore())

    return pd.DataFrame(mytrainingdf)


# lookbacknoofgames=4
# df2018=pd.read_pickle('games2018seasondf.pickle')


# game=Game('NO','ATL',5,df2018)
# game.getmyteamYds('pass',3),game.getmyteamTrnOvr(lookback_noofgames=3)
# game.myteam_allgamesdf

lookbacknoofgames=4
df2018=pd.read_pickle('games2018seasondf.pickle')
df2017=pd.read_pickle('games2017seasondf.pickle')
df2016=pd.read_pickle('games2016seasondf.pickle')
df2015=pd.read_pickle('games2015seasondf.pickle')
df2014=pd.read_pickle('games2014seasondf.pickle')
df2013=pd.read_pickle('games2013seasondf.pickle')
df2012=pd.read_pickle('games2012seasondf.pickle')
df2011=pd.read_pickle('games2011seasondf.pickle')
df2010=pd.read_pickle('games2010seasondf.pickle')
df2009=pd.read_pickle('games2009seasondf.pickle')

trainingset2018=GetTrainingDF(lookbacknoofgames,df2018)
trainingset2017=GetTrainingDF(lookbacknoofgames,df2017)
trainingset2016=GetTrainingDF(lookbacknoofgames,df2016)
trainingset2015=GetTrainingDF(lookbacknoofgames,df2015)
trainingset2014=GetTrainingDF(lookbacknoofgames,df2014)
trainingset2013=GetTrainingDF(lookbacknoofgames,df2013)
trainingset2012=GetTrainingDF(lookbacknoofgames,df2012)
trainingset2011=GetTrainingDF(lookbacknoofgames,df2011)
trainingset2010=GetTrainingDF(lookbacknoofgames,df2010)
trainingset2009=GetTrainingDF(lookbacknoofgames,df2009)

bigtrainingdataset=trainingset2017.append(trainingset2016).append(
    trainingset2015).append(trainingset2014).append(trainingset2013).append(trainingset2012).append(
    trainingset2011)
bigtest2018dataset=trainingset2018

def buildkey(df):
    key=sorted((df['myteam'],df['oppteam']))
    key.append(df['gamedate'].strftime('%Y/%m/%d'))
    return '-'.join(key)

bigtrainingdataset['key']=bigtrainingdataset.apply(buildkey,axis=1)
bigtrainingdataset=bigtrainingdataset.sort_values(by='key').drop_duplicates(subset=['key'], keep='first')

bigtest2018dataset['key']=bigtest2018dataset.apply(buildkey,axis=1)
bigtest2018dataset=bigtest2018dataset.sort_values(by='key').drop_duplicates(subset=['key'], keep='first')

bigtrainingdataset.to_pickle('bigtrainingdataset.pickle')
bigtest2018dataset.to_pickle('bigtest2018dataset.pickle')

len(trainingset2018),len(trainingset2017),len(trainingset2016),len(trainingset2015),len(trainingset2014)

len(trainingset2013),len(trainingset2012),len(trainingset2011),len(trainingset2010),len(trainingset2009)

bigtrainingdataset.head(2).T
