==> Step6_RegressorModel.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split,cross_val_score,GridSearchCV
from sklearn.linear_model import LogisticRegression,RidgeClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.preprocessing import StandardScaler
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.svm import SVC
from sklearn.neural_network import MLPClassifier
%matplotlib inline

rawdf=pd.read_pickle('bigtrainingdataset.pickle')
df=rawdf.drop(['key','myteam','oppteam','gamedate','result'],axis=1)

df.head(2)

def GetAdjustedDF(df,adj_factor):
    
    def AdjustedStats (statlist,adj_factor) :
        if len(statlist)==3:
            return np.matmul(np.array(statlist),np.array([1+adj_factor,1,1-adj_factor]))
        elif len(statlist)==4:
            return np.matmul(np.array(statlist),np.array([1+adj_factor,1+adj_factor/2,1-adj_factor/2,1-adj_factor]))
    
    cols_to_transform=['myteam_prvwins', 'oppteam_prvwins','myteam_scores', 'oppteam_scores', 'myteam_trnovr', 'oppteam_trnovr',
       'myteam_top', 'oppteam_top', 'myteam_pyds', 'oppteam_pyds',
       'myteam_ryds', 'oppteam_ryds', 'myteam_penyds', 'oppteam_penyds',
       'myteam_ptyds', 'oppteam_ptyds']

    adjusteddf=df.copy()
    for col in cols_to_transform:
        adjusteddf[col]=df[col].apply(AdjustedStats,adj_factor=adj_factor)
        
    return adjusteddf

def Get_traintest(df):
    X,y=df.drop(['myteam_curscore','oppteam_curscore','myteam_penyds','oppteam_penyds',
             'myteam_prvwins','oppteam_scores',
#              'myteam_trnovr','oppteam_trnovr','myteam_top','oppteam_top'
            ],axis=1),df[['myteam_curscore','oppteam_curscore']]
    
    X=pd.DataFrame(StandardScaler().fit_transform(X),columns=X.columns)
    
    X_train,X_test,y_train,y_test=train_test_split(X,y,test_size=0.35,shuffle=True)
    return (X_train,X_test,y_train,y_test)

X_train,X_test,y_train,y_test=Get_traintest(GetAdjustedDF(df,adj_factor=1))

from sklearn.linear_model import LinearRegression
regmod=LinearRegression()
regmod.fit(X_train,y_train)
regmod.predict(X_train)

#  Train Test Ready above... Training to follow

mydictofmodels=[
    {
        'estimator':LogisticRegression(),
        'name':'LogisticRegression',
        'parameters':{
            'solver':['lbfgs','liblinear'],
            'C':[1.0,0.8,0.5,0.3]
        }
    },
        {
        'estimator':RidgeClassifier(),
        'name':'Ridge Classifier',
        'parameters':{
            'alpha':[1.0,2.0,10.0,0.5],
            'solver' : ['auto', 'svd', 'cholesky', 'lsqr', 'sparse_cg', 'sag']
        }
    },
        {
        'estimator':GaussianNB(),
        'name':'Gaussian Naive Bayes',
        'parameters':{
#             'solver':['lbfgs','linear']
        }
    },
            {
        'estimator':DecisionTreeClassifier(),
        'name':'Decision Tree',
        'parameters':{
            'min_samples_split':[2,5,10,20,40],
            'min_samples_leaf' :[20,40,60,100],
            'max_depth':[3,5,10,25]
        }
    },
            {
        'estimator':RandomForestClassifier(),
        'name':'Random Forest',
        'parameters':{
            'min_samples_split':[2,5,10,20,40],
            'min_samples_leaf' :[20,40,60,100],
            'n_estimators':[100]
        }
    },
    {
        'estimator':SVC(),
        'name':'SVC',
        'parameters':{
            'gamma':['auto','scale']
        }
    }
    
]

modelnames=[]
modelscores=[]
modelbestparams=[]
adj_factor_list=[]

for adj_factor in [0,0.1,0.2,0.3,0.4,0.5]:
    adjdf=GetAdjustedDF(df=df,adj_factor=adj_factor)
    X_train,X_test,y_train,y_test=Get_traintest(adjdf)
    for model in mydictofmodels:
        gsmodel=GridSearchCV(model['estimator'],param_grid=model['parameters'],cv=5,verbose=True)
        gsmodel.fit(X_train,np.array(y_train).ravel())
        modelnames.append(model['name'])
        modelscores.append(gsmodel.score(X_test,y_test))
        modelbestparams.append(gsmodel.best_params_)
        adj_factor_list.append(adj_factor)


resultsdf=pd.DataFrame({'modelname':modelnames,
              'scores':modelscores,
              'bestparams':modelbestparams,
              'adj_factor':adj_factor_list})
resultsdf.sort_values(['scores'],ascending=False).head(10)

# mlpmodel=MLPClassifier(solver='adam', alpha=1e-5,
#                        hidden_layer_sizes=(5,10,5), early_stopping=False,random_state=1,max_iter=1000,verbose=True,learning_rate_init=0.001)
# mlpmodel.fit(X_train,np.array(y_train).ravel())
# mlpmodel.score(X_train,y_train),mlpmodel.score(X_test,y_test)
